+++ src/initializer.ts
@@ -2,10 +2,9 @@ import * as vscode from 'vscode';
 import { initialize as loaderInit } from './loader';
 import { initialize as commandInit } from './vscode/command';
 
-
 export function initialize(context: vscode.ExtensionContext) {
     // file loader initialize
-    loaderInit();
+    loaderInit(context);
 
     // command register initialize
     commandInit(context);