+++ src/parser.ts
@@ -1,10 +1,16 @@
 // guid:.\w+
 
-export function getGuid(text: string) {
-    let regex = /guid:.\w+/;
-    let match = regex.exec(text);
-    
-    let guid = (match || "").toString();
+export function getGuid(data: string) {
+    let regex = /guid: (.*)/;
+    return data.match(regex)?.[1] ?? "";
+}
 
-    return guid.split(':')[1].trim();
+export function getProjectName(data: string) {
+    let regex = /productName: (.*)/;
+    return data.match(regex)?.[1] ?? "";
+}
+
+export function getProjectVersion(data: string) {
+    let regex = /m_EditorVersionWithRevision: (.*)/;
+    return data.match(regex)?.[1] ?? "";
 }